* Renamed pyfdl.py to fdl to match contents * Switched load(s) and dump(a) to new read_from_file/string and write_to_file/string functions
